+++ fiddles/extjs6/fiddle-0036-ItunesTop100/src/detail-grid.js
@@ -1,7 +1,9 @@
 
 Ext.define('Fiddle.DetailGrid', {
     extend: 'Ext.grid.Panel',
-    store: Ext.data.StoreManager.lookup('tunes'),
+    requires: [
+      'Fiddle.Tunes'
+    ],
     xtype: 'detailgrid',
     columns: [
         {
@@ -38,6 +40,17 @@ Ext.define('Fiddle.DetailGrid', {
             dataIndex: 'link',
             text: 'Link'
         }
-    ]
+    ],
+  constructor: function(config) {
+    let store = Ext.data.StoreManager.lookup('tunes');
+    if (store) {
+      config.store = store;
+    } else {
+      config.store = new Fiddle.Tunes({
+        storeId: 'tunes'
+      });
+    }
+    this.callParent([config]);
+  }
 });