+++ fiddles/extjs6/fiddle-0036-ItunesTop100/src/meta.js
@@ -1,8 +1,6 @@
 var meta = {
-    fiddleHeader: 'Itunes Top 100',
-    fiddleSubHeader: 'Just a simple grid and store.' +
-    '<br />'
+    fiddleHeader: 'Itunes Top 100 Videos',
+    fiddleSubHeader: 'POC that uses ExtJS to serve up the <a href="https://itunes.apple.com/us/rss/topsongs/limit=100/genre=20/explicit=true/xml" target="_blank">itunes top 100 videos rss feed</a>.  It accesses the feed via a CORS POST request to a <a href="http://12-bradyhouse.rhcloud.com/](http://12-bradyhouse.rhcloud.com/">node.js REST Service hosted on OpenShift.com</a>.  See <a href="https://github.com/bradyhouse/house/tree/master/fiddles/node/fiddle-0019-PassThruRest" target="_blank">Node POC #19</a>.' +
+    '<br />',
+    itunesRSSFeed: 'https://itunes.apple.com/us/rss/topmusicvideos/limit=100/genre=1614/explicit=true/json'
 };
-
-
-